Revert "import inst by inst id"
