+++ resources/language/default/property.json
@@ -1,6 +1,5 @@
 {
     "common_property_bk_inst_name":"实例名",
-    "common_property_bk_inst_id":"实例ID",
 
     "biz_property_bk_biz_name": "业务名",
     "biz_property_life_cycle": "生命周期",
@@ -26,7 +25,6 @@
     "module_property_bk_bak_operator": "备份维护人",
     "plat_property_bk_cloud_name": "云区域",
     "plat_property_bk_supplier_account": "供应商",
-    "host_property_bk_host_id": "主机ID",
     "host_property_bk_host_innerip": "内网IP",
     "host_property_bk_host_outerip": "外网IP",
     "host_property_operator": "主要维护人",
