+++ src/scene_server/host_server/host_service/logics/host_add.go
@@ -55,11 +55,6 @@ func AddHost(req *restful.Request, ownerID string, appID int, hostInfos map[int]
 		return err, nil, nil, nil
 	}
 
-	// parase host excel assocate fields
-	for index, host := range hostInfos {
-		hostsInst.ParseHostInstanceAssocate(index, host)
-	}
-
 	var errMsg, updateErrMsg, succMsg []string
 
 	ts := time.Now().UTC()
@@ -85,43 +80,23 @@ func AddHost(req *restful.Request, ownerID string, appID int, hostInfos map[int]
 		if nil == iSubArea {
 			iSubArea = common.BKDefaultDirSubArea
 		}
-
-		var iHostID interface{}
-		var isOK bool
-		if inputType == common.InputTypeExcel {
-			// host not db ,check params host info with host id
-			iHostID, isOK = host[common.BKHostIDField]
-
-		}
-
-		if false == isOK {
-			key := fmt.Sprintf("%s-%v", innerIP, iSubArea)
-			iHost, isDBOK := hostMap[key]
-			if isDBOK {
-				isOK = isDBOK
-				iHostID, _ = iHost[common.BKHostIDField]
-			}
-
-		}
-
-		// delete system fields
-		delete(host, common.BKHostIDField)
+		key := fmt.Sprintf("%s-%v", innerIP, iSubArea)
+		iHost, isOk := hostMap[key]
 
 		//生产日志
-		if isOK {
-
-			hostID, _ := util.GetInt64ByInterface(iHostID)
+		if isOk {
+			hostID, _ := util.GetIntByInterface(iHost[common.BKHostIDField])
 			//prepare the log
 			strHostID := fmt.Sprintf("%d", hostID)
 			logObj := NewHostLog(req, common.BKDefaultOwnerID, strHostID, cc.HostCtrl(), cc.ObjCtrl(), hostLogFields)
 
-			err := hostsInst.UpdateHostInstance(index, host, hostID)
+			err := hostsInst.UpdateHostInstance(index, host, iHost)
 			if nil != err {
 				updateErrMsg = append(updateErrMsg, err.Error())
 				continue
 			}
 			logContent, _ := logObj.GetHostLog(strHostID, false)
-			logConents = append(logConents, auditoplog.AuditLogExt{ID: int(hostID), Content: logContent, ExtKey: innerIP})
+			logConents = append(logConents, auditoplog.AuditLogExt{ID: hostID, Content: logContent, ExtKey: innerIP})
 
 		} else {
 			//prepare the log
@@ -348,37 +323,40 @@ func NewHostsInstance(req *restful.Request, ownerID, inputType string, cloudID i
 	return h, nil
 }
 
-func (h *hostsInstance) ParseHostInstanceAssocate(index int, host map[string]interface{}) {
+func (h *hostsInstance) UpdateHostInstance(index int, host, oldHost map[string]interface{}) error {
 	if common.InputTypeExcel == h.inputType {
-		if _, ok := h.rowErr[index]; true == ok {
-			return
+		if err, ok := h.rowErr[index]; true == ok {
+			return fmt.Errorf(h.langHandle.Languagef("import_row_int_error_str", index, err.Error()))
 		}
 		err := h.assObjectInt.SetObjAsstPropertyVal(host)
 		if nil != err {
 			blog.Error("host assocate property error %d %s", index, err.Error())
-			h.rowErr[index] = err
+			return fmt.Errorf(h.langHandle.Languagef("import_row_int_error_str", index, err.Error()))
 		}
 
 	}
-}
-
-func (h *hostsInstance) UpdateHostInstance(index int, host map[string]interface{}, hostID int64) error {
-
 	//delete(host, common.BKCloudIDField)
 	delete(host, "import_from")
 	delete(host, common.CreateTimeField)
 
+	hostID, _ := util.GetIntByInterface(oldHost[common.BKHostIDField])
+	cloudID, err := util.GetIntByInterface(oldHost[common.BKCloudIDField])
+	if nil != err {
+		blog.Error("get cloud id error %d %s", index, err.Error())
+		return fmt.Errorf(h.langHandle.Languagef("import_row_int_error_str", index, "cloud id "+err.Error()))
+	}
+	host[common.BKCloudIDField] = cloudID
 	filterFields := []string{common.CreateTimeField}
 
 	valid := validator.NewValidMapWithKeyFields(common.BKDefaultOwnerID, common.BKInnerObjIDHost, h.objAddr, filterFields, h.forward, h.errHandle)
-	_, err := valid.ValidMap(host, common.ValidUpdate, int(hostID))
+	_, err = valid.ValidMap(host, common.ValidUpdate, hostID)
 	if nil != err {
 		blog.Error("host valid error %v %v", index, err)
 		return fmt.Errorf(h.langHandle.Languagef("import_row_int_error_str", index, err.Error()))
 
 	}
 	//update host asst attr
-	err = scenecommon.UpdateInstAssociation(h.objAddr, h.req, int(hostID), h.ownerID, common.BKInnerObjIDHost, host) //hostAsstData, ownerID, host)
+	err = scenecommon.UpdateInstAssociation(h.objAddr, h.req, hostID, h.ownerID, common.BKInnerObjIDHost, host) //hostAsstData, ownerID, host)
 	if nil != err {
 		blog.Error("update host asst attr error : %v", err)
 		return fmt.Errorf(h.langHandle.Languagef("import_row_int_error_str", index, err.Error()))
@@ -402,6 +380,17 @@ func (h *hostsInstance) UpdateHostInstance(index int, host map[string]interface{
 }
 
 func (h *hostsInstance) AddHostInstance(index, appID, moduleID int, host map[string]interface{}, ts time.Time) (int, error) {
+	if common.InputTypeExcel == h.inputType {
+		if err, ok := h.rowErr[index]; true == ok {
+			return 0, fmt.Errorf(h.langHandle.Languagef("import_row_int_error_str", index, err.Error()))
+		}
+
+		err := h.assObjectInt.SetObjAsstPropertyVal(host)
+		if nil != err {
+			blog.Error("host assocate property error %v %v", index, err)
+			return 0, fmt.Errorf(h.langHandle.Languagef("import_row_int_error_str", index, err.Error()))
+		}
+	}
 
 	_, ok := host[common.BKCloudIDField]
 	if false == ok {
